+++ libft/ft_memcpy.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ameechan <ameechan@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ameechan <ameechan@student.42.ch>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 19:25:58 by ameechan          #+#    #+#             */
-/*   Updated: 2023/10/11 13:35:56 by ameechan         ###   ########.fr       */
+/*   Updated: 2023/10/19 15:10:49 by ameechan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, int len)
+#include <stdlib.h>
+
+void	*ft_memcpy(void *dest, const void *src, size_t len)
 {
 	unsigned char	*ptr1;
 	unsigned char	*ptr2;
@@ -26,7 +28,7 @@ void	*ft_memcpy(void *dest, const void *src, int len)
 	}
 	return (dest);
 }
-/* 
+/*
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
